requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
anthropic
pyyaml
weasyprint
pytest
httpx
